Sega Genesis: Adding code to draw an image on the screen.

+++ java/net/mikekohn/java_grinder/SegaGenesis.java
@@ -50,7 +50,7 @@ abstract public class SegaGenesis
   public static void setVerticalScroll(int value) { }
 
   /** Set pattern table. */
-  public static void setPatternTable(byte[] patterns) { }
+  public static void setPatternTable(int[] patterns) { }
 
   /** Set image data. */
   public static void setImageData(byte[] image) { }
